Remove c casts in unittest
Bug: 1266991
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4712390
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1174510}

+++ media/filters/hls_data_source_provider_unittest.cc
@@ -79,7 +79,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadAllFromLimitedStream) {
         ASSERT_TRUE(result.has_value());
         auto stream = std::move(result).value();
         ASSERT_FALSE(stream.CanReadMore());
-        ASSERT_EQ(stream.BytesInBuffer(), (size_t)44);
+        ASSERT_EQ(stream.BytesInBuffer(), 44u);
         ASSERT_EQ(std::string(stream.AsStringPiece()),
                   "The Quick Brown Fox Jumped Over The Lazy Dog");
       }));
@@ -112,7 +112,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadDefaultChunkFromLimitedStream) {
         ASSERT_TRUE(result.has_value());
         auto stream = std::move(result).value();
         ASSERT_FALSE(stream.CanReadMore());
-        ASSERT_EQ(stream.BytesInBuffer(), (size_t)44);
+        ASSERT_EQ(stream.BytesInBuffer(), 44u);
         ASSERT_EQ(std::string(stream.AsStringPiece()),
                   "The Quick Brown Fox Jumped Over The Lazy Dog");
       }));
@@ -145,7 +145,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadSmallSizeFromLimitedStream) {
         ASSERT_TRUE(result.has_value());
         auto stream = std::move(result).value();
         ASSERT_TRUE(stream.CanReadMore());
-        ASSERT_EQ(stream.BytesInBuffer(), (size_t)14);
+        ASSERT_EQ(stream.BytesInBuffer(), 14u);
         ASSERT_EQ(std::string(stream.AsStringPiece()), "The Quick Brow");
 
         // Read it again!
@@ -154,7 +154,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadSmallSizeFromLimitedStream) {
               ASSERT_TRUE(result.has_value());
               auto stream = std::move(result).value();
               ASSERT_TRUE(stream.CanReadMore());
-              ASSERT_EQ(stream.BytesInBuffer(), (size_t)28);
+              ASSERT_EQ(stream.BytesInBuffer(), 28u);
               ASSERT_EQ(std::string(stream.AsStringPiece()),
                         "The Quick Brown Fox Jumped O");
             }),
@@ -169,7 +169,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadSmallSizeFromUnlimitedStream) {
         ASSERT_TRUE(result.has_value());
         auto stream = std::move(result).value();
         ASSERT_TRUE(stream.CanReadMore());
-        ASSERT_EQ(stream.BytesInBuffer(), (size_t)14);
+        ASSERT_EQ(stream.BytesInBuffer(), 14u);
         ASSERT_EQ(std::string(stream.AsStringPiece()), "The Quick Brow");
 
         // Read it again!
@@ -178,7 +178,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadSmallSizeFromUnlimitedStream) {
               ASSERT_TRUE(result.has_value());
               auto stream = std::move(result).value();
               ASSERT_TRUE(stream.CanReadMore());
-              ASSERT_EQ(stream.BytesInBuffer(), (size_t)28);
+              ASSERT_EQ(stream.BytesInBuffer(), 28u);
               ASSERT_EQ(std::string(stream.AsStringPiece()),
                         "The Quick Brown Fox Jumped O");
             }),
@@ -193,7 +193,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadSmallSizeWithFlush) {
         ASSERT_TRUE(result.has_value());
         auto stream = std::move(result).value();
         ASSERT_TRUE(stream.CanReadMore());
-        ASSERT_EQ(stream.BytesInBuffer(), (size_t)14);
+        ASSERT_EQ(stream.BytesInBuffer(), 14u);
         ASSERT_EQ(std::string(stream.AsStringPiece()), "The Quick Brow");
 
         // clear the buffer
@@ -205,7 +205,7 @@ TEST(HlsDataSourceStreamUnittest, TestReadSmallSizeWithFlush) {
               ASSERT_TRUE(result.has_value());
               auto stream = std::move(result).value();
               ASSERT_TRUE(stream.CanReadMore());
-              ASSERT_EQ(stream.BytesInBuffer(), (size_t)14);
+              ASSERT_EQ(stream.BytesInBuffer(), 14u);
               ASSERT_EQ(std::string(stream.AsStringPiece()), "n Fox Jumped O");
             }),
             14);
@@ -223,7 +223,7 @@ TEST(HlsDataSourceStreamUnittest, ReadAllFromMultiChunkStream) {
         ASSERT_FALSE(stream.CanReadMore());
         // 0xFFFF   =  65535
         // readsize = 446000
-        ASSERT_EQ(stream.BytesInBuffer(), (size_t)446000);
+        ASSERT_EQ(stream.BytesInBuffer(), 446000u);
       }));
 }
 
@@ -239,7 +239,7 @@ TEST(HlsDataSourceStreamUnittest, ReadAllFromMultiChunkStreamUnknownLength) {
 
         // clear the buffer
         stream.Flush();
-        ASSERT_EQ(stream.BytesInBuffer(), (size_t)0);
+        ASSERT_EQ(stream.BytesInBuffer(), 0u);
 
         std::move(stream).ReadAll(
             base::BindOnce([](HlsDataSourceStream::ReadResult res) {
